+++ source/core/parameters.cc
@@ -1728,51 +1728,70 @@ namespace Parameters
   {
     prm.enter_subsection("FEM");
     {
-      prm.declare_entry("velocity order",
+      prm.declare_entry("velocity degree",
                         "1",
                         Patterns::Integer(),
-                        "interpolation order velocity");
-      prm.declare_entry("pressure order",
+                        "interpolation degree velocity");
+      prm.declare_alias("velocity degree", "velocity order", true);
+      prm.declare_entry("pressure degree",
                         "1",
                         Patterns::Integer(),
-                        "interpolation order pressure");
-      prm.declare_entry("void fraction order",
+                        "interpolation degree pressure");
+      prm.declare_alias("pressure degree", "pressure order", true);
+      prm.declare_entry("void fraction degree",
                         "1",
                         Patterns::Integer(),
-                        "interpolation order void fraction");
-      prm.declare_entry("temperature order",
+                        "interpolation degree void fraction");
+      prm.declare_alias("void fraction degree", "void fraction order", true);
+      prm.declare_entry("temperature degree",
                         "1",
                         Patterns::Integer(),
-                        "interpolation order temperature");
-      prm.declare_entry("tracer order",
+                        "interpolation degree temperature");
+      prm.declare_alias("temperature degree", "temperature order", true);
+      prm.declare_entry("tracer degree",
                         "1",
                         Patterns::Integer(),
-                        "interpolation order tracer");
-      prm.declare_entry("cls order",
+                        "interpolation degree tracer");
+      prm.declare_alias("tracer degree", "tracer order", true);
+      prm.declare_entry("cls degree",
                         "1",
                         Patterns::Integer(),
-                        "interpolation order tracer");
-      prm.declare_alias("cls order", "VOF order", true);
+                        "interpolation degree cls");
+      prm.declare_alias("cls degree", "VOF degree");
+      prm.declare_alias("cls degree", "cls order", true);
+      prm.declare_alias("cls degree", "VOF order", true);
       prm.declare_entry(
-        "phase cahn hilliard order",
+        "phase cahn hilliard degree",
         "1",
         Patterns::Integer(),
-        "interpolation order phase parameter in the Cahn-Hilliard equations");
+        "interpolation degree phase parameter in the Cahn-Hilliard equations");
+      prm.declare_alias("phase cahn hilliard degree",
+                        "phase cahn hilliard order",
+                        true);
       prm.declare_entry(
-        "potential cahn hilliard order",
+        "potential cahn hilliard degree",
         "1",
         Patterns::Integer(),
-        "interpolation order chemical potential in the Cahn-Hilliard equations");
+        "interpolation degree chemical potential in the Cahn-Hilliard equations");
+      prm.declare_alias("potential cahn hilliard degree",
+                        "potential cahn hilliard order",
+                        true);
       prm.declare_entry(
-        "electromagnetics trial order",
+        "electromagnetics trial degree",
         "1",
         Patterns::Integer(),
-        "interpolation order for the trial space of the electromagnetics physics (time-harmonic Maxwell equations).");
+        "interpolation degree for the trial space of the electromagnetics physics (time-harmonic Maxwell equations).");
+      prm.declare_alias("electromagnetics trial degree",
+                        "electromagnetics trial order",
+                        true);
       prm.declare_entry(
-        "electromagnetics test order",
+        "electromagnetics test degree",
         "2",
         Patterns::Integer(),
-        "interpolation order for the test space of the electromagnetics physics (time-harmonic Maxwell equations).");
+        "interpolation degree for the test space of the electromagnetics physics (time-harmonic Maxwell equations).");
+      prm.declare_alias("electromagnetics test degree",
+                        "electromagnetics test order",
+                        true);
 
       prm.declare_entry(
         "tracer uses dg",
@@ -1805,21 +1824,21 @@ namespace Parameters
   {
     prm.enter_subsection("FEM");
     {
-      velocity_order            = prm.get_integer("velocity order");
-      pressure_order            = prm.get_integer("pressure order");
-      void_fraction_order       = prm.get_integer("void fraction order");
-      temperature_order         = prm.get_integer("temperature order");
-      tracer_order              = prm.get_integer("tracer order");
+      velocity_order            = prm.get_integer("velocity degree");
+      pressure_order            = prm.get_integer("pressure degree");
+      void_fraction_order       = prm.get_integer("void fraction degree");
+      temperature_order         = prm.get_integer("temperature degree");
+      tracer_order              = prm.get_integer("tracer degree");
       tracer_uses_dg            = prm.get_bool("tracer uses dg");
-      CLS_order                 = prm.get_integer("cls order");
+      CLS_order                 = prm.get_integer("cls degree");
       CLS_uses_dg               = prm.get_bool("cls uses dg");
-      phase_cahn_hilliard_order = prm.get_integer("phase cahn hilliard order");
+      phase_cahn_hilliard_order = prm.get_integer("phase cahn hilliard degree");
       potential_cahn_hilliard_order =
-        prm.get_integer("potential cahn hilliard order");
+        prm.get_integer("potential cahn hilliard degree");
       electromagnetics_trial_order =
-        prm.get_integer("electromagnetics trial order");
+        prm.get_integer("electromagnetics trial degree");
       electromagnetics_test_order =
-        prm.get_integer("electromagnetics test order");
+        prm.get_integer("electromagnetics test degree");
       enable_bubble_function_velocity =
         prm.get_bool("enable bubble function velocity");
       enable_bubble_function_pressure =